Update PS2Controller.cs
Dont throw an exception on PS/2 Dectection Failure, USB Connected Devices tend to go through a USB to PS/2 Emulator. This means for example mice are PS/2 Devices in this case but dont operate on IRQ12 as the USB Bus Generates a IRQ. so this atleast makes most keyboards work. also the inline can go into the if statement now.