Update VMWareSVGAII.cs
Fix SVGA to avoid a Index out of Bounds crash that happened due to a double memory write.
With this we disable the Driver before writing new values to the FIFO and initiating the memory again, before we initiated the memory regardless if it was already set or not and sent garbage to the FIFO causing it to crash. Also introduce the ability to switch between Graphics And Text mode.