Fixed Nasm Interminable Macro Recursion...
This build of Nasm resolved the tedious error, which is known as an Interminable Macro Recursion. For verification, it was capable of loading a larger byte array without displaying the error. Now, you can add more code to your project without it displaying an error. Just one thing to note: if the asset (or array) that you are importing is too large, VS will start to slow down because of the amount of RAM usage. Also, if (the asset) is way too large VS might restart itself.