AudioMixer: remove obsolete constructor + syntax cleanup
Slight AudioMixer clean-up: this commit makes the syntax cleaner, and removes an obsolete constructor created for an older version of the class when it was still in development.

The removed constructor is not needed, as AudioMixer automatically creates its buffers to compensate for any buffers that will be written to.