Create Console.Cosmos.cs
The cosmos console stream, this serve two purpose, allow to read input from user using a StreamReader (or this stream directly), Allow to write to the System Console using a StreamWriter (or this stream direclty). Cursor Position updates are made on Stream Flush. This may not be the behavior a Stream needs to have, but all changes are made directly instead of waiting to be flushed.